docs(README.md): Update README.md
README.md is not initialised to cover the context of the backend
project.

We need to ensure that everyone working on the backend uses the same
instructions when installing dependencies and debugging the server.

Let's update README.md include context of the backend and continue
updating it as the project progresses.

We should have a consistent and updated README.md so that fellow
developers will have reliable instructions to follow.